Move is_stalemate into chess module
